requirements.txt
# Core LangChain
langchain==0.1.0
langchain-community==0.0.20
langchain-aws==0.1.0

# AWS SDK
boto3>=1.34.72
botocore>=1.34.72

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.23

# OpenSearch
opensearch-py==2.5.0
requests-aws4auth==1.2.3

# Data processing
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0

# Logging and monitoring - SOLO CloudWatch (NO LangSmith)
structlog==24.1.0
watchtower==3.0.1  # CloudWatch Logs handler

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
